+++ selfdrive/ui/sunnypilot/onroad/developer_ui/__init__.py
@@ -11,7 +11,7 @@ from openpilot.selfdrive.ui.sunnypilot.onroad.developer_ui.elements import (
   DesiredLateralAccelElement, ActualLateralAccelElement, DesiredSteeringAngleElement,
   AEgoElement, LeadSpeedElement, FrictionCoefficientElement, LatAccelFactorElement,
   SteeringTorqueEpsElement, BearingDegElement, AltitudeElement, DesiredSteeringPIDElement,
-  LatInterpolElement, LatDeltaElement,
+  LatInterpolElement,
 )
 from openpilot.system.ui.lib.application import gui_app, FontWeight
 from openpilot.system.ui.lib.text_measure import measure_text_cached
@@ -43,7 +43,6 @@ class DeveloperUiRenderer(Widget):
     self.friction_elem = FrictionCoefficientElement()
     self.lat_accel_factor_elem = LatAccelFactorElement()
     self.lat_interp_elem = LatInterpolElement()
-    self.lat_delta_elem = LatDeltaElement()
     self.steering_torque_elem = SteeringTorqueEpsElement()
     self.bearing_elem = BearingDegElement()
     self.altitude_elem = AltitudeElement()
@@ -137,10 +136,9 @@ class DeveloperUiRenderer(Widget):
     elements = []
     is_torque = sm['controlsState'].lateralControlState.which() == 'torqueState'
 
-    # Leftmost: INTERP + ΔCRV (torque only)
+    # Leftmost: INTERP (torque only)
     if is_torque:
       elements.append(self.lat_interp_elem.update(sm, ui_state.is_metric))
-      elements.append(self.lat_delta_elem.update(sm, ui_state.is_metric))
 
     elements.append(self.lead_speed_elem.update(sm, ui_state.is_metric))